backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai>=1.0.0
pandas==2.2.3
python-pptx==1.0.2
python-multipart==0.0.20
pydantic==2.10.6
